Beta
So this is a functionally finished product. There are some leftovers of previous versions and stylistic things to enhance but I am not really adept in good style. 
The function compareWriting(...) is bothering me because of all the arguments, but I was desperate to shorten its parent function deleteUser() which is still like 70 lines long. I think the 30 lines saved by writing compareWriting() were worth it.